Move WaitForPaint helper into a header
No real changes, just make this class, used in wxWindow::SetSize() tests
only so far, reusable in the other tests.
